Refactor ids2nc, extract common tensorization logic in IDSTensorizer
Allows reuse of functionality to (partially) convert IDSs to xarray Datasets.
